first sketch of the instructions decoding process
Just added a structure that represents a RISC-V instruction with
enough data so we can differenciate it from others when decoding
the instruction. The idea is that doing (word & ix.data) will give
us a unique identifier of the instruction such that we can compare
it against ix.result to know whenever it is that instruction or not.

This probably needs a cache on top cause looping `INSTRUCTIONS` all
the time doesn't seem too efficient, but we'll see. :)